refactor: created new hooks
* Created Hooks for ingredient modals and recipe calculations.
